default

+++ img_rename.py
@@ -10,8 +10,9 @@ sorted_file_lst = sorted(file_name_and_time_lst, key=lambda x: x[1], reverse=Tru
 recent_file = sorted_file_lst[0]
 start_int = int(input('숫자를 입력해주세요 : '))
 for image in sorted_file_lst:
-    src = os.path.join(files_path, image[0])
-    dst = str(start_int) + '.png'
-    dst = os.path.join(files_path, dst)
-    os.rename(src, dst)
-    start_int += 1
+    if image[0] != ".gitkeep":
+        src = os.path.join(files_path, image[0])
+        dst = str(start_int) + '.png'
+        dst = os.path.join(files_path, dst)
+        os.rename(src, dst)
+        start_int += 1
